New feature: HelpQueueFeature
* Purpose:
This feature is designed to aid in the process of maintaining a help
queue list, as would normally be found in a classroom where students
sign their name on the white board when asking for help by the teacher.

The feature is using only a few commands where students ask the bot for
help and they will automatically be queued.

* Restrictions
The only person(s) that can dequeue people are teachers. The name
"teacher" as the role, is hard-coded in to the feature. this requires
the user on the server that uses the dequeuing command to be teachers.
Students cannot perform this action, and they cannot register more than
once in the queue.

The teacher cannot dequeue by using PM, since roles need to be parsed
and these cannot be seen outside of the server channels.